+++ javascript/client/src/types/tables.ts
@@ -2,5 +2,10 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export * from './files'
-export * from './tables'
+
+export interface Header {
+	name: string
+	fieldName: string
+	iconName?: string
+	width: number
+}
