+++ javascript/client/src/utils/index.ts
@@ -15,3 +15,4 @@ export * from './lang'
 export * from './noop'
 export * from './sessionStorage'
 export * from './stats'
+export * from './tables'
